+++ src/components/List.jsx
@@ -3,17 +3,14 @@ import {
   collection,
   onSnapshot,
   Timestamp,
-  doc,
-  updateDoc,
   query,
   orderBy,
-  deleteDoc,
 } from 'firebase/firestore';
 import { db } from '../lib/firebase';
 import { getUser } from '../storage-utils/storage-utils';
 import Nav from './Nav';
 import { useNavigate } from 'react-router-dom';
-import { calculateEstimate } from '@the-collab-lab/shopping-list-utils';
+import ListItem from './ListItem';
 
 export default function List() {
   //states:
@@ -21,28 +18,6 @@ export default function List() {
   const [userToken] = useState(getUser());
   const navigate = useNavigate();
   const [searchInputValue, setSearchInputValue] = useState('');
-
-  const checkboxHandler = (e, item) => {
-    e.preventDefault();
-
-    const docItem = doc(db, userToken, item.id);
-    let now = Timestamp.now().seconds;
-    const totalPurchases = item.data().totalPurchases + 1;
-
-    const daysSinceLastTransaction = daysSinceLastPurchase(item);
-
-    const estimatedPurchaseInterval = calculateEstimate(
-      item.data().estimatedPurchaseInterval,
-      daysSinceLastTransaction,
-      totalPurchases,
-    );
-
-    updateDoc(docItem, {
-      lastPurchaseDate: now,
-      totalPurchases: totalPurchases,
-      estimatedPurchaseInterval: estimatedPurchaseInterval,
-    });
-  };
 
   const wasPurchasedWithin24Hours = (item) => {
     let now = Timestamp.now().seconds;
@@ -75,20 +50,6 @@ export default function List() {
     );
   };
 
-  const determinePurchaseCategory = (item) => {
-    if (!isActive(item)) {
-      return 'inactive';
-    }
-
-    if (daysUntilNextPurchase(item) < 7) {
-      return 'soon';
-    } else if (daysUntilNextPurchase(item) <= 30) {
-      return 'kinda-soon';
-    } else {
-      return 'not-soon';
-    }
-  };
-
   const sortList = (docs) => {
     docs.sort(function (a, b) {
       if ((isActive(a) && isActive(b)) || (!isActive(a) && !isActive(b))) {
@@ -117,17 +78,6 @@ export default function List() {
       unsubscribe();
     };
   }, [userToken]);
-
-  const deleteHandler = (item) => {
-    const deletionConfirmation = window.confirm(
-      `Are you sure you'd like to delete ${
-        item.data().item
-      } from your shopping list?`,
-    );
-    if (deletionConfirmation) {
-      deleteDoc(doc(db, userToken, item.id));
-    }
-  };
 
   return (
     <>
@@ -166,31 +116,7 @@ export default function List() {
               })
 
               .map((item, index) => {
-                return (
-                  <div>
-                    <li key={index}>
-                      <label
-                        className={determinePurchaseCategory(item)}
-                        aria-label={`next purchase is ${determinePurchaseCategory(
-                          item,
-                        )}`}
-                      >
-                        <input
-                          aria-label="checkbox for purchased item"
-                          id={item.data().id}
-                          type="checkbox"
-                          onChange={(e) => checkboxHandler(e, item)}
-                          checked={wasPurchasedWithin24Hours(item)}
-                          disabled={wasPurchasedWithin24Hours(item)}
-                        />
-                      </label>
-                      {item.data().item}
-                      <button onClick={() => deleteHandler(item)}>
-                        delete
-                      </button>
-                    </li>
-                  </div>
-                );
+                return <ListItem item={item} index={index} />;
               })}
           </ul>
         </div>
